% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell24}
\alias{cell24}
\title{A data.frame of 24 immune cells name from Bindea et al}
\format{A data.frame with 24 rows and 2 column}
\usage{
cell24
}
\description{
It's a built-in data. The first column represents the abbreviation of 24 immune cells, the second column represents the full name of 24 immune cells
}
\references{
Bindea G, Mlecnik B, Tosolini M, Kirilovsky A, Waldner M, Obenauf AC, et al. Spatiotemporal dynamics of intratumoral immune cells reveal the immune landscape in human cancer. Immunity. 2013;39:782–95.
}
\keyword{datasets}
