% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell64}
\alias{cell64}
\title{A data.frame of 64 immune cells name from xCell method}
\format{A data.frame with 64 rows and 2 column}
\usage{
cell64
}
\description{
It's a built-in data. The first column represents the abbreviation of 64 immune cells, the second column represents the full name of 64 immune cells
}
\references{
Aran D , Hu Z , Butte A J . xCell: digitally portraying the tissue cellular heterogeneity landscape[J]. Genome Biology, 2017, 18(1):220.
}
\keyword{datasets}
