\name{envData}
\alias{envData}
\alias{exp.example}
\alias{cellmatrix}
\alias{mutcell}
\alias{mutmatrix}
\docType{data}

\title{
envData
}
\description{
The variables in the environment include an example expression profiles, a cell abundance matrix, a binary numerical matrix which shows the immune cells driven by somatic mutation, a binary mutations matrix.
}
\details{
The environment variable includes the variable \code{exp.example}, \code{cellmatrix},\code{mutcell},\code{mutmatrix}
}

\format{An environment variable}
\author{Junwei Han\email{hanjunwei1981@163.com},Baotong Zheng\email{btzheng1116@163.com}}

\keyword{data}
