% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExampleData.R
\name{GetExampleData}
\alias{GetExampleData}
\title{Get the example data}
\usage{
GetExampleData(exampleData)
}
\arguments{
\item{exampleData}{A character, should be one of "exp.example", "cellmatrix", "mutcell", "mutmatrix", "surv".}
}
\description{
Get the example data from SMDIC package.
}
\details{
The function `GetExampleData(ExampleData = "mutmatrix)")` obtains the mutations matrix
}
\references{
Subramanian, A., Tamayo, P., Mootha, V.K., Mukherjee, S., Ebert, B.L., Gillette, M.A., Paulovich, A., Pomeroy, S.L., Golub, T.R., Lander, E.S. et al. (2005) Gene set enrichment analysis: a knowledgebased approach for interpreting genome-wide expression profiles. Proc Natl Acad Sci U S A, 102, 15545-15550.
}
