% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immunelist}
\alias{immunelist}
\title{A large list of 24 immune cells type-specific gene signatures from Bindea et al}
\format{A list}
\usage{
immunelist
}
\description{
It's a built-in data. The name of the list represent 24 immune cells, the value of the list are 24 immune cells type-specific gene signatures from Bindea et al
}
\references{
Bindea G, Mlecnik B, Tosolini M, Kirilovsky A, Waldner M, Obenauf AC, et al. Spatiotemporal dynamics of intratumoral immune cells reveal the immune landscape in human cancer. Immunity. 2013;39:782–95.
}
\keyword{datasets}
