% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotwaterfall}
\alias{plotwaterfall}
\title{plotwaterfall}
\usage{
plotwaterfall(maffile, mutcell.summary, cellnumcuoff = 4)
}
\arguments{
\item{maffile}{The name of mutation annotation file (MAF) format data. It must be an absolute path or the name  relatived to the current working directory.}

\item{mutcell.summary}{The result of `mutcellsummary` function}

\item{cellnumcuoff}{a threshold value (4 as the default value). The mutation genes which drive at least "cellnumcuoff" cells are retained for drawing an waterfall.}
}
\description{
Function‘plotwaterfall‘ plots the waterfall for mutation genes which drive immune cells.
}
\examples{
file<-"dir" #dir must be an absolute path or the name  relatived to the current working directory.
\dontrun{plotwaterfall(maffile = dir,mutcell.summary = summary,cellnumcuoff =0)}
}
