% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMDIC-package.R
\docType{package}
\name{SMDIC-package}
\alias{SMDIC-package}
\title{Identification of somatic mutation-driven immune cells}
\description{
With the use of functions in this packages, users could identify the immune cells driven by somatic mutations in tumor microenvironment.
}
