% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmapcell}
\alias{heatmapcell}
\title{heatmapcell}
\usage{
heatmapcell(gene, mutcell, cellmatrix, mutmatrix)
}
\arguments{
\item{gene}{Somatic mutant gene name}

\item{mutcell}{A list, mutcell is the result of function `mutcorcell`.}

\item{cellmatrix}{Cell abundance matrix, cellmatrix is the result of function `exp2cell`.}

\item{mutmatrix}{A binary mutations matrix, in which 1 represents any mutation occurs in a particular gene in a particular sample, otherwise the element is 0.}
}
\description{
A function to draw clustered heatmaps for the cells driven by a somatic mutation.
}
\examples{
mutcell<-GetExampleData("mutcell") # The result of `mutcorcell` function.
cellmatrix<-GetExampleData("cellmatrix") # Cell abundance matrix
mutmatrix<-GetExampleData("mutmatrix") # A binary mutations matrix
heatmapcell(gene = "TP53",mutcell = mutcell,cellmatrix = cellmatrix,mutmatrix = mutmatrix)
}
